Made small edit to the outputs. Changed the P and N outputs from the timer so it is possible to use the timer the way it is meant as well.
Timer1 (and 8) on the STM32F4 supports 3 phase PWM with complenetary outputs out of the box. It will handle all the signals, and the firmware only needs to handle commutation events. This is a great advantage when using standard PWM schemes, nad will probably be the final solution. However, the Chibios driver does not support this, so I must make a custom driver for it. This is also a good reason for doing it the brute force way I'm doing now. Once the functionality is tested, and most of the bugs are weeded out, I can start to utilize the timer better.